+++ configureWebpack.js
@@ -527,6 +527,11 @@ async function configureWebpack(env) {
                                                         '@fortawesome/pro-solid-svg-icons/${member}',
                                                     skipDefaultConversion: true
                                                 },
+                                                '@fortawesome/pro-thin-svg-icons': {
+                                                    transform:
+                                                        '@fortawesome/pro-thin-svg-icons/${member}',
+                                                    skipDefaultConversion: true
+                                                },
                                                 '@fortawesome/free-brands-svg-icons': {
                                                     transform:
                                                         '@fortawesome/free-brands-svg-icons/${member}',