+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * Updated special transformations for the FontAwesome icon dependencies to include their "thin" package, which was added
   to hoist-react last year and was not being properly tree-shaken.
+* Updated Babel configuration to enable key transforms via preset-env `includes` directive. Resolves issue with outdated
+  plugin names in the prior config causing errors like "cannot find package @babel/plugin-proposal-class-properties"
+  for apps that updated to Babel 7.22 (several key transforms were released under new names with the 7.22 release -
+  see https://github.com/babel/babel/pull/15614).
 
 ### 📚 Libraries